+++ src/Shared/Avalon.Common/ValueObjects/CharacterId.cs
@@ -1,11 +1,11 @@
 namespace Avalon.Common.ValueObjects;
 
-public class CharacterId : ValueObject<ulong>, IHideObjectMembers
+public class CharacterId : ValueObject<uint>, IHideObjectMembers
 {
-    public CharacterId(ulong value) : base(value)
+    public CharacterId(uint value) : base(value)
     {
     }
     
-    public static implicit operator ulong(CharacterId characterId) => characterId.Value;
-    public static implicit operator CharacterId(ulong value) => new CharacterId(value);
+    public static implicit operator uint(CharacterId characterId) => characterId.Value;
+    public static implicit operator CharacterId(uint value) => new CharacterId(value);
 }
